+++ src/ska_low_mccs_spshw/tile/tile_device.py
@@ -5740,13 +5740,13 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
     pointingDelays = attribute_from_signal(
         pointingDelays_signal,
         dtype=(("DevFloat",),),
-        max_dim_x=32,  # Channels
-        max_dim_y=8,  # Antennas
+        max_dim_x=32,  # delay, delay rate for each of the 16 antennas
+        max_dim_y=48,  # Beams
         rel_change=0.01,
         archive_rel_change=0.01,
         write_to_signal=True,
         label="Pointing Delays",
-        doc="Get the pointing delays for beam 0.",
+        doc="Get the pointing delays for all beams.",
     )
 
     currentDraw = attribute_from_signal(
@@ -6805,7 +6805,7 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         Specify the delay in seconds and the delay rate in seconds/second.
 
         The delay_array specifies the delay and delay rate for each antenna. beam_index
-        specifies which beam is desired (range 0-7)
+        specifies which beam is desired (range 0-47)
 
         :param argin: An array containing: beam index,
             the delay in seconds and the delay rate in
@@ -6835,7 +6835,7 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
             self.logger.error("Insufficient parameters")
             raise ValueError("Insufficient parameters")
         beam_index = int(argin[0])
-        if beam_index < 0 or beam_index > 7:
+        if beam_index < 0 or beam_index > 47:
             self.logger.error("Invalid beam index")
             raise ValueError("Invalid beam index")
         delay_array = []