+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased
 
 * [THORN-650] Convert tile attributes to be signal backed. Add test to read/write/readback attribute values.
+* [THORN-671] Bump tpm firmware version used when bios version > 1.0.0 from 11.0.0 -> 12.0.0.
+This comes with some 48 beam support. Removing 8 beam limitations and adding visibility of all 48 beams pointingDelays.
 * [SKB-1328] Restrict the VerifyEvents property to archive events only. Change events are now always pushed (detect=False).
 * [LOW-2454] Improve error detection, handling and logging during SpsStation's initialise sequence. All MccsTile command failures should now cause initialisation to fail, some critical operations are retried, and initialisation steps are logged at INFO level rather than DEBUG.
 * [THORN-698] Fix AcquireDataForCalibration when DAQ is running.