+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Version History
 
+## Unreleased
+* [SKB-767] Update the minimum integration_time for Integrated LMC Data
+
 ## 15.0.0
 
 * [SPRTS-1067] Don't reset tile ref time on subrack events, reset on actual TPM connection.