+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Version History
 
 ## Unreleased
+* [SKB-1482] Update `tpm-api` from **4.0.1** to **4.0.2**.
 * [SKB-767] Update the minimum integration_time for Integrated LMC Data
 
 ## 15.0.0