+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Version History
 
-## Unreleased
+## 15.0.0
 
 * [THORN-650] Convert tile attributes to be signal backed. Add test to read/write/readback attribute values. Removes usage of AttributeManager in MccsTile.
+* [SPRTS-1067] Don't reset tile ref time on subrack events, reset on actual TPM connection.
 * [THORN-671] Bump tpm firmware version used when bios version > 1.0.0 from 11.0.0 -> 12.0.0.
 This comes with some 48 beam support. Removing 8 beam limitations and adding visibility of all 48 beams pointingDelays.
 * [SKB-1328] Restrict the VerifyEvents property to archive events only. Change events are now always pushed (detect=False).